Fix sierra camlib compilation warnings.
* library.c:
  * sierra_list_folders(...): Change 'i' type to signed for trivial 'for' loop where it is compared to signed integer.
  * sierra_read_packet(...): Unsigned 16-bit little-endian checksum is compiled from two 8-bit values, but the bytes are not cast to correct size. Fix the issue by casting bytes to larger unsigned values.
  * sierra_set_speed(...): Add cast to 'unsigned int' to 'settings.serial.speed' in signed-unsigned compare. This retains current code behavior.
* olympus-desc.c: Comment out unused variables.
* sierra-desc.c:
  * camera_cam_desc_get_widget(...): Change 'value' variable type to 'unsigned int'. Previously this variable pointer was already cast to unsigned for filling and used in signed-unsigned compare as unsigned. Hence it has no behavior changes.
  * All other changes are basically the same. Signed 'for' loop counters were changed to unsigned type as they were already used as unsigned integers in loop comparisons.
* sierra-usbwrap.c: Cast signed integer to unsigned to quiet signed-unsigned comparison warning. No behavior change.
* sierra.c:
  * Comment out unused static variables.
  * put_file_func(...): Ensure 'available_memory' is not negative and cast to unsigned int to quiet compiler warning.
  * camera_init(...): Cast 'i' to unsigned int to quiet signed-unsigned compare warning. No behavior change.

+++ camlibs/sierra/sierra-usbwrap.c
@@ -459,7 +459,7 @@ usb_wrap_DATA (GPPort *dev, unsigned int type, char *sierra_response, int *sierr
    msg_len = msg_len * 256 + (unsigned int)(size.c2);
    msg_len = msg_len * 256 + (unsigned int)(size.c1);
 
-   if (*sierra_len < msg_len - sizeof(*msg))
+   if ((unsigned int)*sierra_len < msg_len - sizeof(*msg))
    {
       GP_DEBUG( "usb_wrap_read_packet buffer too small! (%i < %i) *** FAILED", *sierra_len, msg_len);
       return GP_ERROR;
